refactor(research): extract EPUB PDF evidence validation
